% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dual_ascent_adapt}
\alias{dual_ascent_adapt}
\title{Solve the projected dual ascent problem with fixed weights and adaptive step size}
\usage{
dual_ascent_adapt(X, Phi, weights, Lambda0, maxiter, eps, nv0, trace)
}
\arguments{
\item{X}{the data, with the columns being units, the rows being features}

\item{Phi}{the edge incidence matrix, defined as Phi_{li} = 1 if(l_1 == i); -1 if(l_2 == i); 0 otherwise}

\item{weights}{the non-zero weights in a vector}

\item{Lambda0}{the initial guess of Lambda}

\item{maxiter}{maximum iterations}

\item{eps}{the duality gap tolerence}

\item{trace}{whether save the primal and dual values of every iteration}

\item{nv}{initial step size}
}
\value{
a list including U, V, Lambda and number of iterations to convergence
}
\description{
Solve the projected dual ascent problem with fixed weights and adaptive step size
}
