% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_solve.R
\name{fusion_cluster_path}
\alias{fusion_cluster_path}
\title{Function to calculate the clustering path}
\usage{
fusion_cluster_path(X, lambda, gamma, nnn, tol.nn = 0, maxiter.mm = 200,
  maxiter.cvx = 30000, tol = 1e-05)
}
\arguments{
\item{X}{the data, with the columns being units, the rows being features}

\item{lambda}{a vector of the tuning parameter}

\item{gamma}{a scaler of the MCP tuning parameter}

\item{nnn}{number of nearest neighbors for the preprocessing}

\item{tol.nn}{nearest neighbor tolerance for the preprocessing}
}
\value{
the list of the centroids and the list their differences, for each lambda
}
\description{
Function to calculate the clustering path
}
