% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvxclustr.R
\docType{package}
\name{ncvxclustr}
\alias{ncvxclustr}
\alias{ncvxclustr-package}
\title{ncvxclustr: non-convex clustering}
\description{
algorithm to solve the clustering problem by non-convex penalization
}
\section{Functions}{

}

