% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_visual.R
\name{plot_path}
\alias{plot_path}
\title{Function to plot the clustering path}
\usage{
plot_path(X, path)
}
\arguments{
\item{X}{the data, with the columns being units, the rows being features}

\item{path}{the output of function \link{fusion_cluster_path}}
}
\description{
Function to plot the clustering path
}
